GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - feat(README): Update
 (refs/gitbutler/feat-README-Update-)
   branch head: 8c67f71bc088dfb257c2862936bb1e814693253d
   - notes/homebrew.md
   - brew.sh
   - .gitignore
   - README.md

Your previous branch was: refs/heads/main

The sha for that commit was: 742903a45e568928d0549cc0a7788c2716691757

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch